Get FreeBSD compiling again
Check for explicit be32enc and be32dec functions with configure probe and add
to the endian.h compat header if necessary; do not check for FreeBSD or not.
