+++ src/compat/endian.h
@@ -191,6 +191,26 @@ inline uint64_t le64toh(uint64_t little_endian_64bits)
 }
 #endif // HAVE_DECL_LE64TOH
 
+#if HAVE_DECL_BE32DEC == 0
+static inline uint32_t be32dec(const void *pp)
+{
+	const uint8_t *p = (uint8_t const *)pp;
+	return ((uint32_t)(p[3]) + ((uint32_t)(p[2]) << 8) +
+	    ((uint32_t)(p[1]) << 16) + ((uint32_t)(p[0]) << 24));
+}
+#endif // HAVE_DECL_BE32DEC
+
+#if HAVE_DECL_BE32ENC == 0
+static inline void be32enc(void *pp, uint32_t x)
+{
+	uint8_t *p = (uint8_t *)pp;
+	p[3] = x & 0xff;
+	p[2] = (x >> 8) & 0xff;
+	p[1] = (x >> 16) & 0xff;
+	p[0] = (x >> 24) & 0xff;
+}
+#endif // HAVE_DECL_BE32ENC
+
 #endif // WORDS_BIGENDIAN
 
 #endif // BITCOIN_COMPAT_ENDIAN_H
